requirements.txt
Flask~=2.0.3
PyGithub~=1.55